Clean up common functions.
